WIP: Create os-nosdn-osm scenario
The tasks added by this change
- configure opnfv host by installing required OSM packages, configuring
lxd and lxd bridge, making devuser member of lxd and docker groups
- clone OSM devops repo and checks out specific version
- execute install_osm.sh script with certain arguments to ensure script
runs unattendedo

RFC:
A new environment variable, SCENARIO_POST_DEPLOY, is introduced to control
when the role should be executed by XCI. It can be controller by setting
this variable in xci_overrides and the default value is false.

TODO:
- add tasks to register VIM which is installed by os-nosdn-nofeature
scenario

FUTURE:
- contact OSM and OPNFV Functest project to determing first set of test
cases which could be the deployment of Cirros based network service. [1]

[1] https://osm.etsi.org/wikipub/index.php/OSM_Release_FOUR#Deploying_your_first_Network_Service

Change-Id: I8897288f7d0cadc4931ca21de13a4f4418fff7a8
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>