Create os-nosdn-osm scenario
The tasks added by this change
- configure OPNFV host by installing required OSM packages, configuring
lxd and lxd bridge, making devuser member of lxd and docker groups
- clone OSM devops repo and check out specific version
- execute install_osm.sh script with certain arguments to ensure script
runs unattendedo

An important thing regarding this scenario and its role is that no OSM
specific task exists as part of pre-deployment and everything is done as
post-deployment. In order to achieve that, the role's main.yml contains
debug statement to log this info and skip/move on. Everything related to
OSM and the scenario itself is done in post-deployment.yml which is a
way to run post-deployment tasks for this type of scenarios. A new playbook
is also created in releng-xci/xci/installer/osa/playbooks with the same
name which checks the existence of scenario post-deployment.yml and runs
the tasks defined in it. If the file doesn't exist, post-deployment is
skipped.

[1] https://osm.etsi.org/wikipub/index.php/OSM_Release_FOUR#Deploying_your_first_Network_Service

Change-Id: I8897288f7d0cadc4931ca21de13a4f4418fff7a8
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>