Register OpenStack as VIM to OSM
This patch creates osmrc file and registers the
OpenStack as VIM to OSM.

TODO:
Currently we don't have any suitable way to find out
the functioning status of OSM as soon it was installed.
So we pause the execution of playbook for 2 minutes and
then register OpenStack as VIM. As soon as healthchecks
are added to Docker containers(upstream), we'll use them
to identify the functioning status of OSM and modify the
existing task.

installer-type:osa
deploy-scenario:os-nosdn-osm

Change-Id: Ic38bd2c142f8b1c19bff125f08bab198c423fe5d
Signed-off-by: Venkata Harshavardhan Reddy Allu <venkataharshavardhan_ven@srmuniv.edu.in>