Register OpenStack as VIM to OSM
This patch creates osmrc file and registers the
OpenStack as VIM to OSM. To register OS as VIM,
OSM needs access to OS credentials, auth url which
are provided from the other patch in releng-xci:

https://gerrit.opnfv.org/gerrit/#/c/63135/

This patch works in combination with the above patch
which hadn't been merged to releng-xci repo yet, but
was tested at a different patch:

https://gerrit.opnfv.org/gerrit/#/c/63161/

TODO:
Currently we don't have any suitable way to find out
the functioning status of OSM as soon it was installed.
So we pause the execution of playbook for 2 minutes and
then register OpenStack as VIM. As soon as healthchecks
were added to Docker containers(upstream), we'll use them
to identify the functioning status of OSM and modify the
existing task.

installer-type:osa
deploy-scenario:os-nosdn-osm

Change-Id: Ic38bd2c142f8b1c19bff125f08bab198c423fe5d
Signed-off-by: Venkata Harshavardhan Reddy Allu <venkataharshavardhan_ven@srmuniv.edu.in>