osm: Drop configure-opnfvhost.yml playbook
With the recent changes in th OSM(upstream), OSM can now configure
and setup LXD unattended during the installation. So, we don't need
the 'configure-opnfvhost.yaml' playbook anymore. This commit also
modifies a task to wait for the OSM containers to be ready using
grep rather than a simple timeout.

Change-Id: Ic09b85594b64e4c1db3fbdc9a2dc7a382802d787
Signed-off-by: Venkata Harshavardhan Reddy Allu <venkataharshavardhan_ven@srmuniv.edu.in>