Enable the use of multiple installers per scenario
A new installer, OpenStack-Helm is being introduced to XCI and this
requires the XCI Framework to have capability to use the right installer
for given scenario.

This change updates the directory structure for the scenarios and
adds a new folder named after the specific installer that supports the
scenario.

A corresponding change will be sent to releng-xci to adapt the framework.

installer-type:kubespray
deploy-scenario:k8-calico-nofeature

Change-Id: Iea85a7a6b81fe516502d3922733c69875465f267
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>